Changed directory check code in Decompress
I changed this as Unity kept throwing an exception "ArgumentException: Destination is not a directory. You need to specify the name of a directory, instead". The problem arises because GetAttributes does a bitwise combination of all attributes so a compressed directory would not have an attribute value equal to a directory. Instead a "HasFlag" check gets the job done.